pyproject.toml
[tool.poetry]
name = "bookstore"
version = "1.2.0"
description = "Web application of a bookstore intended for the sale of paper books online."
authors = ["Aliaksandr Dauhuchyts <alexander.dovguchits@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.3"
Django = "4.1.1"
asgiref = "3.5.2"
python-dotenv = "1.0.0"
crispy-bootstrap5 = "0.7"
django-crispy-forms = "1.14.0"
Pillow = "9.2.0"
sqlparse = "0.4.3"
tzdata = "2022.4"
trycourier = "4.4.0"
social-auth-app-django = "5.0.0"
social-auth-core = "4.3.0"
psycopg2-binary = "^2.9.9"
dj-database-url = "^2.1.0"
pytest = "^7.4.3"
pytest-django = "^4.7.0"


[tool.poetry.group.dev.dependencies]
autoflake = "^2.1.1"
black = "^23.3.0"
django-stubs = "^4.2.0"
flake8 = "^6.0.0"
flake8-annotations-complexity = "^0.0.7"
flake8-bandit = "^4.1.1"
flake8-black = "^0.3.6"
flake8-bugbear = "^23.5.9"
flake8-builtins = "^2.1.0"
flake8-cognitive-complexity = "^0.1.0"
flake8-comprehensions = "^3.12.0"
flake8-debugger = "^4.1.2"
flake8-executable = "^2.1.3"
flake8-expression-complexity = "^0.0.11"
flake8-fixme = "^1.1.1"
flake8-multiline-containers = "^0.0.19"
flake8-mutable = "^1.2.0"
flake8-pep3101 = "^2.0.0"
flake8-print = "^5.0.0"
flake8-printf-formatting = "^1.1.2"
flake8-quotes = "^3.3.2"
flake8-simplify = "^0.20.0"
flake8-variables-names = "^0.0.5"
httpie = "^3.2.2"
ipython = "^8.13.2"
mypy = "^1.3.0"
pep8-naming = "^0.13.3"
pyproject-flake8 = "^6.0.0.post1"
types-requests = "^2.30.0.0"
pytest = "^7.3.1"
flake8-import-order = "^0.18.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
cache_dir = ".local/mypy/"
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_calls = false
disallow_untyped_defs = false
exclude = [".local",".git",".idea",".vscode",]
files = ["."]
no_implicit_optional = true
plugins = [
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main",
]
pretty = true
python_version = "3.11"
show_error_context = true
sqlite_cache = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
ignore_missing_imports = true
module = [
    "djoser.*",
    "pyodbc.*",
]

[tool.django-stubs]
django_settings_module = "project.settings"

[tool.black]
line-length = 120
target-version = ["py311"]
include = '^.+\.pyi?$'
extend-exclude = '''
(
    .*\.local.*
    | .*\.idea.*
    | .*\.vscode.*
    | .*\.git.*
)
'''
